Initial multi-vendor support for Flex
