+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/dividat/driver
 
-go 1.12
+go 1.22
 
 require (
 	github.com/cenkalti/backoff v2.2.1+incompatible
@@ -17,8 +17,16 @@ require (
 	// Both projects are dormant at the moment, but we might want to re-evaluate this
 	// dependency choice as these projects evolve in the future.
 	github.com/libp2p/zeroconf/v2 v2.2.0
-
 	github.com/pin/tftp v2.1.0+incompatible
 	github.com/sirupsen/logrus v1.8.1
 	go.bug.st/serial v1.6.1
 )
+
+require (
+	github.com/creack/goselect v0.1.2 // indirect
+	github.com/miekg/dns v1.1.43 // indirect
+	golang.org/x/net v0.0.0-20210423184538-5f58ad60dda6 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+)
+
+replace go.bug.st/serial => github.com/yfyf/go-serial v1.6.4-withproduct
